refactor: Move fork_bomb from func to module Co-Authored-By: Idan Kalmanzon idankalmanzon@gmail.com
